commit comment for last time (oops) reworked InterpreterTester into several pieces: first, analogues to the RFCSystem classes: SimSystem.cs, a FieldState.cs that acts as IInfoAcceptor and IPredictor, a SimController.cs IController.
The SimEngine takes care of stepping the simulation, and the FieldView takes care of drawing

I fixed the debug assertion by adding a bunch of locks and returning copies of the IPredictor arrays.

Todo: a RefboxState class, clean up the arrow display (probably into some sort of Overlay decorator class), automated refereeing =D and allowing the display to be resized

Still not as clean as it could be, mostly because the arrow display.