Fixed an old bug with orientation: Math.Atan2 returns either a positive or a negative angle, but predictor expects value not to flip sign from frame to frame.
Moved the check of "is ball on top of robot" outside the loop for patterned robots, so that it applies to patternless robots as well.

Changed to look for a ball blob of maximum area, not with an area closest to some predefined value.

Added more strict checking of whether a blob is a center dot (check area width, height, and width/height ratio). The check is in isCenterDot function