Assorted improvements to the AveragingPredictor.
Most important: weighing the average by time since last seen does not work because the update from one of the cameras is always much much fresher than the one from the other. Currently, just takes the arithmetic average.

Added various checks to make sure robots cannot appear on top of each other etc.

Added the weighted average between the measurements and old values in update: this was supposed to be there in the first place.