Altered constants for short range PID, each tuned per robot, for the sake of kick planning.
Tuned kick planning constants, most notably the distance and difference in orientation that is required for getting to the point behind the ball.

Changed constants for forward veer planner to improve going forward towards the ball in the process of kicking. These cause it to hit the ball more often, though many kicks that would otherwise have been a miss are now rather inaccurate kicks. These constants deserve future consideration.