The constants for motion planning and kicking as they stand at the end of the US Open competition. There has been extensive calibration. Notably there were many changes to the PID constants for short range motion control, though the optimal values of these constants tend to change over time, particularly with battery life.
In kickplanning, adjusted the precision required for the point behind the ball, the distance to position behind the ball, and the saturation limit for how many successive iterations are needed to recognize the robot as being in the proper direction.

In motionplanning, adjusted the speed and PID constants for the robot driving forward through the ball, and added PID constants for robot spinning.