Added a SSL Server to the SSL wrappers.
The code changes in client code are mysterious issues with the win C++ compiler/linker: had to construct/destruct Net::UDP object explicitly because otherwise (if declared as member in the user class) the "auto-generated" code in constructor for the user class needed the destructor for Net::UDP (why? for some sort of failure path probably) and the linker couldn't find it (although it should). It is not clear why this was not an issue before. Whatever.