The Singapore version of the various constants.
In default.txt, added a NUM_ROBOTS constant used in quite a few parts of the code.

In kickplanning.txt, decreased the lateral distance that a robot could be from the "start kicking" point from .04 to .03 meters. Experience indicated it improved the robot's ability to kick consistently, especially given our more accurate driving and the rule preventing double touching of the ball.

In motionplanning.txt, added constants for the linear scaling of speeds given by ModelFeedback. Each robot ID has an individual scaling factor that can be calibrated, and SPEED_SCALING_FACTOR_ALL is used to scale all at once. Also adjusted the size of the field to Singapore size, which is commented in this commit so that the code will work in Maxwell Dworkin with zero modifications.