Plays as they are in Singapore.
In DEFENSE_2vs0.txt, tightens up the spacing of the two man wall from 22 cm radius to 10 cm radius- infinitely more effective in games. In DEFENSE_3vs0.txt, also tightened the spacing of the front two men, and added the case where the goalie should try to kick the ball.

In OFFENSE_1vs0.txt, made sure the offender wouldn't chase the ball into the goal zone if there was already a robot there- avoids double defending.

PKICK THEIRS MAINPLAY.txt used to assign plays that weren't existent, with disastrous results! For a quick fix during a time out, I turned it into a play that just froze all the robots where they were. This is obviously not a good solution, and I am putting it on my ToDo list for future CS.