Added the PlaySystem project, which is a prototype implementation of an entirely new strategy system (with plays written in C#). See PlaySystem at the wiki for details.
Added PlaySystem to AllProjects.sln, and a reference to it to Play Selector/PlaySelector.csproj.

Integrated the PlaySystem into Interpreter.cs. A constant in default.txt, USE_C_SHARP_PLAY_SYSTEM, decides whether or not to use the new play system (when it is set to false, the PlaySystem project should have no effect on the rest of the code.

Fixed a bug in FlipPredictor where it didn't pass on the team to the flipped Robot Info. I don't know why this didn't cause problems in the old play system (another bug down the line canceling it out?) but it is problematic in PlaySystem.

Added constants to default.txt and plays.txt to work with PlaySystem.

Fixed an entirely unrelated bug in ModelFeedback.cs, left over from a Singapore commit, that caused all robots not to move.